# docs/swagger.yaml
definitions:
...
  user-management_internal_models.CreateUserRequest:
    properties:
      department:
        type: string
      email:
        type: string
      name:
        type: string
      nik:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      position:
        type: string
      role:
        enum:
        - admin
        - supervisor
        - warehouse_operator
        - viewer
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
...
  user-management_internal_models.UpdateUserRequest:
    properties:
      department:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      nik:
        type: string
      phone:
        type: string
      position:
        type: string
      role:
        enum:
        - admin
        - supervisor
        - warehouse_operator
        - viewer
        type: string
      username:
        type: string
    type: object
  user-management_internal_models.UserResponse:
    properties:
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      name:
        type: string
      nik:
        type: string
      phone:
        type: string
      position:
        type: string
      role:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  user-management_internal_utils.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/users:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by email or name
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user-management_internal_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/user-management_internal_models.UserResponse'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user-management_internal_models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Create user
      tags:
      - Users
  /api/v1/users/{userId}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete user (soft delete)
      tags:
      - Users
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user-management_internal_models.UpdateUserRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
...
swagger: "2.0"